all tests pass(kinda), without cp

+++ signals.cpp
@@ -22,22 +22,26 @@ void ctrlZHandler(int sig_num) { //todo: check with pipe and redirections
             else{
                 JobsList::JobEntry* pipeJob1 = smash.jobsList->getJobByPid(pipePid1);
                 pipeJob1->state = STOPPED;
-                smash.jobsList->changeJobId(pipeJob1, smash.jobsList->maxJobID + 1);
-                smash.jobsList->maxJobID++;
+                if (pipeJob1->formerJobId == NOVALUE) {
+                    smash.jobsList->changeJobId(pipeJob1, smash.jobsList->maxJobID + 1);
+                    smash.jobsList->maxJobID++;
+                } else {
+                    smash.jobsList->changeJobId(pipeJob1, pipeJob1->formerJobId);
+                }
                 cout << "smash: process " << pipePid1<< " was stopped" << endl;
             }
         }
-        else{
-            JobsList::JobEntry* pipeJob1 = smash.jobsList->getJobByPid(pipePid1);
-            pipeJob1->state = STOPPED;
-            if (pipeJob1->formerJobId == NOVALUE){
-                smash.jobsList->changeJobId(pipeJob1, smash.jobsList->maxJobID + 1);
-                smash.jobsList->maxJobID++;
-            } else {
-                smash.jobsList->changeJobId(pipeJob1, pipeJob1->formerJobId);
-            }
-            cout << "smash: process " << pipePid1<< " was stopped" << endl;
-        }
+//        else{ // todo delete
+//            JobsList::JobEntry* pipeJob1 = smash.jobsList->getJobByPid(pipePid1);
+//            pipeJob1->state = STOPPED;
+//            if (pipeJob1->formerJobId == NOVALUE){
+//                smash.jobsList->changeJobId(pipeJob1, smash.jobsList->maxJobID + 1);
+//                smash.jobsList->maxJobID++;
+//            } else {
+//                smash.jobsList->changeJobId(pipeJob1, pipeJob1->formerJobId);
+//            }
+//            cout << "smash: process " << pipePid1<< " was stopped" << endl;
+//        }
         if (kill(pipePid2, SIGSTOP) != 0){
             perror("smash error: kill failed"); // kill is the syscall that failed
             return;
@@ -54,6 +58,7 @@ void ctrlZHandler(int sig_num) { //todo: check with pipe and redirections
 
             cout << "smash: process " << pipePid2 << " was stopped" << endl;
         }
+        smash.jobsList->hasPipeInFg = false;
         return;
     }
     JobsList::JobEntry* fgJob = smash.jobsList->getFgJob();
@@ -85,22 +90,27 @@ void ctrlCHandler(int sig_num) { //todo: check with pipe and redirections
     if (smash.jobsList->hasPipeInFg){
         pid_t pipePid1 = smash.jobsList->pipePid1;
         pid_t pipePid2 = smash.jobsList->pipePid2;
-        if (kill(pipePid1, SIGKILL) != 0){
-            perror("smash error: kill failed"); // kill is the syscall that failed
-            return;
+        if (pipePid1 != 0){
+            if (kill(pipePid1, SIGKILL) != 0){
+                perror("smash error: kill failed"); // kill is the syscall that failed
+                return;
+            }
+            else{
+                smash.jobsList->removeJobByPid(pipePid1);
+                cout << "smash: process " << pipePid1 << " was killed" << endl;
+            }
         }
-        else{
-            smash.jobsList->removeJobByPid(pipePid1);
-            cout << "smash: process " << pipePid1 << " was killed" << endl;
+        if (pipePid2 != 0){
+            if (kill(pipePid2, SIGKILL) != 0){
+                perror("smash error: kill failed"); // kill is the syscall that failed
+                return;
+            }
+            else{
+                smash.jobsList->removeJobByPid(pipePid2);
+                cout << "smash: process " << pipePid2 << " was killed" << endl;
+            }
         }
-        if (kill(pipePid2, SIGKILL) != 0){
-            perror("smash error: kill failed"); // kill is the syscall that failed
-            return;
-        }
-        else{
-            smash.jobsList->removeJobByPid(pipePid2);
-            cout << "smash: process " << pipePid2 << " was killed" << endl;
-        }
+        smash.jobsList->hasPipeInFg = false;
         return;
     }
     JobsList::JobEntry* fgJob = smash.jobsList->getFgJob();
@@ -126,10 +136,12 @@ void alarmHandler(int sig_num) {
   // TODO: Add your implementation
     cout << "smash: got an alarm" << endl;
     SmallShell& smash = SmallShell::getInstance();
-//    smash.jobsList->_printTimeoutVector(); //todo: debug
+//    smash.jobsList->removeFinishedJobs();
     time_t* tempTime= NULL;
     time_t now = time(tempTime);
     for (int i = 0; i < smash.jobsList->timeoutJobs.size(); i++){
+       // cout << "i is: " << i << endl; //todo: debug
+//        smash.jobsList->_printTimeoutVector(); //todo: debug
         JobsList::JobEntry* job;
 //        cout << "cond 1: " << (smash.jobsList->timeoutJobs[i].pipe == PIPE1) << endl; // todo: debug
 //        cout << "cond 2: " << (smash.jobsList->timeoutJobs[i].pipe == PIPE2) << endl; // todo: debug
@@ -144,20 +156,29 @@ void alarmHandler(int sig_num) {
         else {
             job = smash.jobsList->getJobById(smash.jobsList->timeoutJobs[i].id);
         }
+
+
         if (job == NULL){
 //            cout << "job is null" << endl;  // todo: debug
-            smash.jobsList->removeTimeoutJob(smash.jobsList->timeoutJobs[i].id);
+//            smash.jobsList->removeTimeoutJob(smash.jobsList->timeoutJobs[i].id);
+            continue;
         }
-        if (now - job->timeStamp >= smash.jobsList->timeoutJobs[i].sleepTime){
-            if (kill(job->command->getPID(), SIGKILL) != 0){
-                perror("smash error: kill failed"); // kill is the syscall that failed
-                return;
-            }
-            else{
-                string cmd_line1 = job->command->cmd_line;
-                cout << "smash: process " << cmd_line1 << " timed out!" << endl;
+        if (job->command->getPID() != smash.smashPid){
+//            cout << "timeout job id: " << smash.jobsList->timeoutJobs[i].id << " sleep: " << smash.jobsList->timeoutJobs[i].sleepTime << endl; //todo: debug
+//            cout << "job id: " << job->jobID << " job cmd: " << job->command->cmd_line << endl;//todo: debug
+            int sleepTime = smash.jobsList->timeoutJobs[i].sleepTime;
+            if (now - job->timeStamp >= smash.jobsList->timeoutJobs[i].sleepTime){
+                if (kill(job->command->getPID(), SIGKILL) != 0){
+                    perror("smash error: kill failed"); // kill is the syscall that failed
+                    return;
+                }
+                else{
+                    string cmd_line1 = job->command->cmd_line;
+                    cout << "smash: timeout " << sleepTime << " " << cmd_line1 << " timed out!" << endl;
+                    smash.jobsList->removeTimeoutJob(smash.jobsList->timeoutJobs[i].id, smash.jobsList->timeoutJobs[i].pipe);
 //                smash.jobsList->removeFinishedJobs(); // todo :check
-//                i--;// todo :check
+                    i--;
+                }
             }
         }
     }
